+++ test/integration/braintree/transaction_gateway_us_bank_account_spec.js
@@ -290,8 +290,7 @@ describe("TransactionGateway", function () {
     let gateway = specHelper.merchant2Gateway;
 
     context("plaid verified", function () {
-      // skipping until CI is more stable
-      it.skip("transacts successfully", function (done) {
+      it("transacts successfully", function (done) {
         specHelper.generatePlaidUsBankAccountNonce(gateway, function (nonce) {
           let transactionParams = {
             merchantAccountId: MerchantAccountTest.AnotherUsBankMerchantAccount,