+++ test/unit/braintree/credit_card_verification_gateway_spec.js
@@ -3,6 +3,8 @@
 let CreditCardVerificationGateway =
   require("../../../lib/braintree/credit_card_verification_gateway").CreditCardVerificationGateway;
 let errorTypes = require("../../../lib/braintree/error_types").errorTypes;
+let ThreeDSecurePassThruNetwork =
+  require("../../../lib/braintree/three_d_secure_pass_thru_network").ThreeDSecurePassThruNetwork;
 
 describe("CreditCardVerificationGateway", () =>
   describe("create", function () {
@@ -62,6 +64,34 @@ describe("CreditCardVerificationGateway", () =>
         assert.isNull(err);
         assert.exists(result);
         assert.deepEqual(result.options.accountInformationInquiry, "send_data");
+        done();
+      });
+    });
+
+    it("accepts threeDSecurePassThru network", function (done) {
+      let verificationGateway = new CreditCardVerificationGateway(fakeGateway);
+      let verificationParams = {
+        creditCard: {
+          number: "4111111111111111",
+          expirationDate: "05/2029",
+        },
+        threeDSecurePassThru: {
+          eciFlag: "05",
+          cavv: "some_cavv",
+          xid: "some_xid",
+          threeDSecureVersion: "2.2.0",
+          authenticationResponse: "Y",
+          directoryResponse: "Y",
+          cavvAlgorithm: "2",
+          dsTransactionId: "some_ds_id",
+          network: ThreeDSecurePassThruNetwork.Visa,
+        },
+      };
+
+      verificationGateway.create(verificationParams, (err, params) => {
+        assert.notExists(err);
+        assert.exists(params);
+        assert.equal(params.verification.threeDSecurePassThru.network, "Visa");
         done();
       });
     });